+++ services/mariadb/wait/wait_test.go
@@ -21,11 +21,7 @@ type apiClientInstanceMocked struct {
 	resourceDescription        string
 }
 
-var (
-	instanceTypeCreate = InstanceTypeCreate
-	instanceTypeUpdate = InstanceTypeUpdate
-	instanceTypeDelete = InstanceTypeDelete
-)
+const deleteOperation = "delete"
 
 func (a *apiClientInstanceMocked) GetInstanceExecute(_ context.Context, _, _ string) (*mariadb.Instance, error) {
 	if a.getFails {
@@ -33,14 +29,13 @@ func (a *apiClientInstanceMocked) GetInstanceExecute(_ context.Context, _, _ str
 			StatusCode: 500,
 		}
 	}
-	if *a.resourceOperation == InstanceTypeDelete && a.resourceState == InstanceStateSuccess {
+	if a.resourceOperation != nil && *a.resourceOperation == deleteOperation && a.resourceState == InstanceStatusActive {
 		if a.deletionSucceedsWithErrors {
 			return &mariadb.Instance{
 				InstanceId: &a.resourceId,
+				Status:     &a.resourceState,
 				LastOperation: &mariadb.InstanceLastOperation{
 					Description: &a.resourceDescription,
-					Type:        a.resourceOperation,
-					State:       &a.resourceState,
 				},
 			}, nil
 		}
@@ -51,11 +46,7 @@ func (a *apiClientInstanceMocked) GetInstanceExecute(_ context.Context, _, _ str
 
 	return &mariadb.Instance{
 		InstanceId: &a.resourceId,
-		LastOperation: &mariadb.InstanceLastOperation{
-			Description: &a.resourceDescription,
-			Type:        a.resourceOperation,
-			State:       &a.resourceState,
-		},
+		Status:     &a.resourceState,
 	}, nil
 }
 
@@ -96,14 +87,14 @@ func TestCreateInstanceWaitHandler(t *testing.T) {
 		{
 			desc:          "create_succeeded",
 			getFails:      false,
-			resourceState: InstanceStateSuccess,
+			resourceState: InstanceStatusActive,
 			wantErr:       false,
 			wantResp:      true,
 		},
 		{
 			desc:          "create_failed",
 			getFails:      false,
-			resourceState: InstanceStateFailed,
+			resourceState: InstanceStatusFailed,
 			wantErr:       true,
 			wantResp:      true,
 		},
@@ -126,21 +117,16 @@ func TestCreateInstanceWaitHandler(t *testing.T) {
 			instanceId := "foo-bar"
 
 			apiClient := &apiClientInstanceMocked{
-				getFails:          tt.getFails,
-				resourceId:        instanceId,
-				resourceOperation: &instanceTypeCreate,
-				resourceState:     tt.resourceState,
+				getFails:      tt.getFails,
+				resourceId:    instanceId,
+				resourceState: tt.resourceState,
 			}
 
 			var wantRes *mariadb.Instance
 			if tt.wantResp {
 				wantRes = &mariadb.Instance{
 					InstanceId: &instanceId,
-					LastOperation: &mariadb.InstanceLastOperation{
-						Type:        &instanceTypeCreate,
-						State:       utils.Ptr(tt.resourceState),
-						Description: utils.Ptr(""),
-					},
+					Status:     utils.Ptr(tt.resourceState),
 				}
 			}
 
@@ -170,14 +156,14 @@ func TestUpdateInstanceWaitHandler(t *testing.T) {
 		{
 			desc:          "update_succeeded",
 			getFails:      false,
-			resourceState: InstanceStateSuccess,
+			resourceState: InstanceStatusActive,
 			wantErr:       false,
 			wantResp:      true,
 		},
 		{
 			desc:          "update_failed",
 			getFails:      false,
-			resourceState: InstanceStateFailed,
+			resourceState: InstanceStatusFailed,
 			wantErr:       true,
 			wantResp:      true,
 		},
@@ -200,21 +186,16 @@ func TestUpdateInstanceWaitHandler(t *testing.T) {
 			instanceId := "foo-bar"
 
 			apiClient := &apiClientInstanceMocked{
-				getFails:          tt.getFails,
-				resourceId:        instanceId,
-				resourceOperation: &instanceTypeUpdate,
-				resourceState:     tt.resourceState,
+				getFails:      tt.getFails,
+				resourceId:    instanceId,
+				resourceState: tt.resourceState,
 			}
 
 			var wantRes *mariadb.Instance
 			if tt.wantResp {
 				wantRes = &mariadb.Instance{
 					InstanceId: &instanceId,
-					LastOperation: &mariadb.InstanceLastOperation{
-						Type:        &instanceTypeUpdate,
-						State:       utils.Ptr(tt.resourceState),
-						Description: utils.Ptr(""),
-					},
+					Status:     utils.Ptr(tt.resourceState),
 				}
 			}
 
@@ -245,20 +226,20 @@ func TestDeleteInstanceWaitHandler(t *testing.T) {
 			desc:                      "delete_succeeded",
 			getFails:                  false,
 			deleteSucceeedsWithErrors: false,
-			resourceState:             InstanceStateSuccess,
+			resourceState:             InstanceStatusActive,
 			wantErr:                   false,
 		},
 		{
 			desc:                      "delete_failed",
 			getFails:                  false,
 			deleteSucceeedsWithErrors: false,
-			resourceState:             InstanceStateFailed,
+			resourceState:             InstanceStatusFailed,
 			wantErr:                   true,
 		},
 		{
 			desc:                      "delete_succeeds_with_errors",
 			getFails:                  false,
-			resourceState:             InstanceStateSuccess,
+			resourceState:             InstanceStatusActive,
 			deleteSucceeedsWithErrors: true,
 			resourceDescription:       "Deleting resource: cf failed with error: DeleteFailed",
 			wantErr:                   true,
@@ -278,7 +259,7 @@ func TestDeleteInstanceWaitHandler(t *testing.T) {
 				getFails:                   tt.getFails,
 				deletionSucceedsWithErrors: tt.deleteSucceeedsWithErrors,
 				resourceId:                 instanceId,
-				resourceOperation:          &instanceTypeDelete,
+				resourceOperation:          utils.Ptr(deleteOperation),
 				resourceDescription:        tt.resourceDescription,
 				resourceState:              tt.resourceState,
 			}