+++ services/opensearch/CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.20.2 (2025-04-02)
+- **Bugfix:** `PartialUpdateInstanceWaitHandler` does not finish when update is succeeded
+- **Deprecation:** Deprecated `InstanceStateSuccess`, `InstanceStateFailed`, `InstanceTypeCreate`, `InstanceTypeUpdate`, `InstanceTypeDelete` and will be removed after 2nd October 2025
+
 ## v0.20.1 (2025-03-19)
 - **Internal:** Backwards compatible change to generated code